Updated test modules

+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Application/Customers/DeleteCustomer/DeleteCustomerCommandHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using AzureFunctions.TestApplication.Domain.Common.Exceptions;
 using AzureFunctions.TestApplication.Domain.Repositories;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
@@ -25,6 +26,11 @@ namespace AzureFunctions.TestApplication.Application.Customers.DeleteCustomer
         public async Task<Unit> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
             var existingCustomer = await _customerRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (existingCustomer is null)
+            {
+                throw new NotFoundException($"Could not find Customer {request.Id}");
+            }
             _customerRepository.Remove(existingCustomer);
             return Unit.Value;
         }
