+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Application/Implementation/SampleDomainsService.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using AzureFunctions.TestApplication.Application.Interfaces;
 using AzureFunctions.TestApplication.Application.SampleDomains;
+using AzureFunctions.TestApplication.Domain.Common.Exceptions;
 using AzureFunctions.TestApplication.Domain.Entities;
 using AzureFunctions.TestApplication.Domain.Repositories;
 using Intent.RoslynWeaver.Attributes;
@@ -38,21 +39,26 @@ namespace AzureFunctions.TestApplication.Application.Implementation
                 Attribute = dto.Attribute,
             };
             _sampleDomainRepository.Add(newSampleDomain);
-            await _sampleDomainRepository.UnitOfWork.SaveChangesAsync();
+            await _sampleDomainRepository.UnitOfWork.SaveChangesAsync(cancellationToken);
             return newSampleDomain.Id;
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<SampleDomainDto> FindSampleDomainById(Guid id, CancellationToken cancellationToken = default)
         {
-            var element = await _sampleDomainRepository.FindByIdAsync(id);
+            var element = await _sampleDomainRepository.FindByIdAsync(id, cancellationToken);
+
+            if (element is null)
+            {
+                throw new NotFoundException($"Could not find SampleDomain {id}");
+            }
             return element.MapToSampleDomainDto(_mapper);
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<List<SampleDomainDto>> FindSampleDomains(CancellationToken cancellationToken = default)
         {
-            var elements = await _sampleDomainRepository.FindAllAsync();
+            var elements = await _sampleDomainRepository.FindAllAsync(cancellationToken);
             return elements.MapToSampleDomainDtoList(_mapper);
         }
 
@@ -62,14 +68,24 @@ namespace AzureFunctions.TestApplication.Application.Implementation
             SampleDomainUpdateDto dto,
             CancellationToken cancellationToken = default)
         {
-            var existingSampleDomain = await _sampleDomainRepository.FindByIdAsync(id);
+            var existingSampleDomain = await _sampleDomainRepository.FindByIdAsync(id, cancellationToken);
+
+            if (existingSampleDomain is null)
+            {
+                throw new NotFoundException($"Could not find SampleDomain {id}");
+            }
             existingSampleDomain.Attribute = dto.Attribute;
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task DeleteSampleDomain(Guid id, CancellationToken cancellationToken = default)
         {
-            var existingSampleDomain = await _sampleDomainRepository.FindByIdAsync(id);
+            var existingSampleDomain = await _sampleDomainRepository.FindByIdAsync(id, cancellationToken);
+
+            if (existingSampleDomain is null)
+            {
+                throw new NotFoundException($"Could not find SampleDomain {id}");
+            }
             _sampleDomainRepository.Remove(existingSampleDomain);
         }
