+++ Tests/CosmosDB/CosmosDB.Application/ClassContainers/DeleteClassContainer/DeleteClassContainerCommandHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using CosmosDB.Domain.Common.Exceptions;
 using CosmosDB.Domain.Repositories;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
@@ -25,6 +26,11 @@ namespace CosmosDB.Application.ClassContainers.DeleteClassContainer
         public async Task<Unit> Handle(DeleteClassContainerCommand request, CancellationToken cancellationToken)
         {
             var existingClassContainer = await _classContainerRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (existingClassContainer is null)
+            {
+                throw new NotFoundException($"Could not find ClassContainer {request.Id}");
+            }
             _classContainerRepository.Remove(existingClassContainer);
             return Unit.Value;
         }
