+++ Tests/GraphQL.CQRS.TestApplication/GraphQL.CQRS.TestApplication.Application/Implementation/ProductsService.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using GraphQL.CQRS.TestApplication.Application.Interfaces;
 using GraphQL.CQRS.TestApplication.Application.Products;
+using GraphQL.CQRS.TestApplication.Domain.Common.Exceptions;
 using GraphQL.CQRS.TestApplication.Domain.Entities;
 using GraphQL.CQRS.TestApplication.Domain.Repositories;
 using Intent.RoslynWeaver.Attributes;
@@ -38,21 +39,26 @@ namespace GraphQL.CQRS.TestApplication.Application.Implementation
                 IsActive = dto.IsActive,
             };
             _productRepository.Add(newProduct);
-            await _productRepository.UnitOfWork.SaveChangesAsync();
+            await _productRepository.UnitOfWork.SaveChangesAsync(cancellationToken);
             return newProduct.MapToProductDto(_mapper);
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<ProductDto> FindProductById(Guid id, CancellationToken cancellationToken = default)
         {
-            var element = await _productRepository.FindByIdAsync(id);
+            var element = await _productRepository.FindByIdAsync(id, cancellationToken);
+
+            if (element is null)
+            {
+                throw new NotFoundException($"Could not find Product {id}");
+            }
             return element.MapToProductDto(_mapper);
         }
 
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<List<ProductDto>> FindProducts(CancellationToken cancellationToken = default)
         {
-            var elements = await _productRepository.FindAllAsync();
+            var elements = await _productRepository.FindAllAsync(cancellationToken);
             return elements.MapToProductDtoList(_mapper);
         }
 
@@ -62,7 +68,12 @@ namespace GraphQL.CQRS.TestApplication.Application.Implementation
             ProductUpdateDto dto,
             CancellationToken cancellationToken = default)
         {
-            var existingProduct = await _productRepository.FindByIdAsync(id);
+            var existingProduct = await _productRepository.FindByIdAsync(id, cancellationToken);
+
+            if (existingProduct is null)
+            {
+                throw new NotFoundException($"Could not find Product {id}");
+            }
             existingProduct.Name = dto.Name;
             existingProduct.Description = dto.Description;
             existingProduct.IsActive = dto.IsActive;
@@ -72,7 +83,12 @@ namespace GraphQL.CQRS.TestApplication.Application.Implementation
         [IntentManaged(Mode.Fully, Body = Mode.Fully)]
         public async Task<ProductDto> DeleteProduct(Guid id, CancellationToken cancellationToken = default)
         {
-            var existingProduct = await _productRepository.FindByIdAsync(id);
+            var existingProduct = await _productRepository.FindByIdAsync(id, cancellationToken);
+
+            if (existingProduct is null)
+            {
+                throw new NotFoundException($"Could not find Product {id}");
+            }
             _productRepository.Remove(existingProduct);
             return existingProduct.MapToProductDto(_mapper);
         }
