+++ Tests/GraphQL.MongoDb.TestApplication/GraphQL.MongoDb.TestApplication.Application/Users/DeleteUser/DeleteUserCommandHandler.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
+using GraphQL.MongoDb.TestApplication.Domain.Common.Exceptions;
 using GraphQL.MongoDb.TestApplication.Domain.Repositories;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
@@ -28,6 +29,11 @@ namespace GraphQL.MongoDb.TestApplication.Application.Users.DeleteUser
         public async Task<UserDto> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
             var existingUser = await _userRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (existingUser is null)
+            {
+                throw new NotFoundException($"Could not find User {request.Id}");
+            }
             _userRepository.Remove(existingUser);
             return existingUser.MapToUserDto(_mapper);
         }
