+++ Tests/Publish.CleanArchDapr.TestApplication/Publish.CleanArchDapr.TestApplication.Application/Customers/DeleteCustomer/DeleteCustomerCommandHandler.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
+using Publish.CleanArchDapr.TestApplication.Domain.Common.Exceptions;
 using Publish.CleanArchDapr.TestApplication.Domain.Repositories;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -25,6 +26,11 @@ namespace Publish.CleanArchDapr.TestApplication.Application.Customers.DeleteCust
         public async Task<Unit> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
         {
             var existingCustomer = await _customerRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (existingCustomer is null)
+            {
+                throw new NotFoundException($"Could not find Customer {request.Id}");
+            }
             _customerRepository.Remove(existingCustomer);
             return Unit.Value;
         }
